try to get axios call to work
