+++ packages/studio/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "baseUrl": ".",
-    "rootDir": ".",
+    "rootDir": "../",
     "outDir": "dist",
     "module": "esnext",
     "target": "es5",
